% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_exec.R
\name{code_exec}
\alias{code_exec}
\title{Run a code chunk at the command line}
\usage{
code_exec(code_path, code_block, file_prefix, folder_output)
}
\arguments{
\item{code_path}{Character. Path to an R script file to be executed.}

\item{code_block}{Character vector. R code to be executed.}

\item{file_prefix}{Character. Prefix for the output files.}

\item{folder_output}{Character. Path to the folder where output files will be saved.}
}
\description{
# This function allows you to run a block of R code or an R script file at the command line,
}
\details{
This function executes R code at the command line. Code can be input
  either as a path to a script using \code{code_path} or as a quoted code block using
  \code{code_block}. The output is then read and returned as a character vector.
}
\examples{
\dontrun{
out_temp <- tempdir()

# Using code_block
code_exec(code_block="Sys.info()", file_prefix="sysinfo", folder_output=out_temp)

# Using code_path (pre-existing R script)
code_exec(code_path=file.path(out_temp, "sysinfo.R"), file_prefix="sysinfo", folder_output=out_temp)
}

}
