% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqualify_setup.R
\name{rqualify_setup}
\alias{rqualify_setup}
\title{Initialize environment for IQ-OQ.}
\usage{
rqualify_setup(setup_tinytex = TRUE, setup_pandoc = TRUE, verbose = TRUE)
}
\arguments{
\item{setup_tinytex}{Logical. If TRUE, sets up TinyTeX for LaTeX document
generation. Note, this does not install the tinytex R package, but the TinyTeX 
LaTeX bundle. It is a convenient wrapper for installing TinyTeX using 
`tinytex::install_tinytex()`, and adding the TinyTeX location to the environment
The function installes the "TinyTeX" bundle and the additional package `grfext`,
and sets the TinyTeX installation path on the system PATH.}

\item{setup_pandoc}{Logical. If TRUE, sets up pandoc for document conversion.
Note, this does not install the pandoc R package, but the Pandoc software. It
is a convenient wrapper around `pandoc::pandoc_install()` and `pandoc::pandoc_activate()`, 
which are called internally.}

\item{verbose}{Logical. If TRUE, prints progress messages to the console.}
}
\value{
A list with the installed Pandoc and TinyTeX versions, though the primary
 purpose of this function is its side effects, setting up TinyTeX and Pandoc as specified.
}
\description{
Initialize environment for IQ-OQ.
}
\details{
`rqualify_setup` allows users to easily install TinyTeX and Pandoc
  as required for rendering Rmd files to PDF. The process can
  be quite time consuming, especially if TinyTeX needs to be installed, and there
  may be apparent hanging processes -- this is expected behavior.
}
\examples{
rqualify_setup(setup_tinytex = FALSE,
               setup_pandoc  = FALSE)

}
\seealso{
\code{\link{rqualify}} which calls this function internally.
}
