% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqualify.R
\name{rqualify}
\alias{rqualify}
\title{Run IQ-OQ on an installation of R software}
\usage{
rqualify(
  path_save,
  setup_tinytex = TRUE,
  setup_pandoc = TRUE,
  file_rmd = "R-validation.Rmd",
  verbose = TRUE
)
}
\arguments{
\item{path_save}{Character. Path to save the R-validation folder. Ensure
a folder named `R-validation` does not already exist at this location.}

\item{setup_tinytex}{Logical. If TRUE, sets up tinytex for LaTeX document
generation.}

\item{setup_pandoc}{Logical. If TRUE, sets up pandoc for document conversion.}

\item{file_rmd}{String. Name of RMarkdown file to render. Default is "R-validation.Rmd". 
The RMarkdown file should be located in the `inst/qualify_r` folder of the package.}

\item{verbose}{Logical. If TRUE, prints progress messages to the console.}
}
\value{
This function does not return a value. It performs side effects by 
  rendering a RMarkdown document.
}
\description{
Run IQ-OQ on an installation of R software
}
\details{
This function creates a folder named R-validation at the specified path,
sets up the environment, and executes the validation process.
The output of the validation will be saved in the created folder. 

The validation process involves running a series of tests on the R installation and
can be quite time consuming. The function will print progress messages to the 
console if `verbose` is set to TRUE.

The following steps are carried out:

\enumerate{
  \item Create the folder tree R-validation/IQ-OQ-TestOutput at path_save, or selected folder
  \item Install tinytex and necessary LaTeX packages
  \item Install pandoc
  \item Copy R-validation.Rmd to the R-validation folder
  \item Execute the IQ-OQ by rendering R-validation.Rmd to R-validation.tex
  \item Compile R-validation.tex to R-validation.pdf
}
}
\examples{
\dontshow{if (tinytex::is_tinytex() && pandoc::pandoc_available()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\donttest{
rqualify(path_save     = tempdir(),
         setup_tinytex = FALSE,
         setup_pandoc  = FALSE,
         file_rmd      = "R-validationPreamble.Rmd")
}
\dontshow{\}) # examplesIf}
}
