% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqualify.R
\name{rqualify}
\alias{rqualify}
\title{Run IQ-OQ on an installation of R software}
\usage{
rqualify(
  path_save,
  setup_tinytex = TRUE,
  setup_pandoc = TRUE,
  file_rmd = "R-validation.Rmd",
  verbose = TRUE
)
}
\arguments{
\item{path_save}{Character. Path to save the R-validation folder. Ensure
a folder named `R-validation` does not already exist at this location.}

\item{setup_tinytex}{Logical. If TRUE, sets up TinyTeX for LaTeX document
generation. Note, this does not install the tinytex R package, but the TinyTeX 
LaTeX bundle. It is a convenient wrapper for installing TinyTeX using 
`tinytex::install_tinytex()`, and adding the TinyTeX location to the environment
path. See \code{\link{rqualify_setup}} for more details.}

\item{setup_pandoc}{Logical. If TRUE, sets up pandoc for document conversion.
Note, this does not install the pandoc R package, but the Pandoc software. It
is a convenient wrapper around `pandoc::pandoc_install()` and `pandoc::pandoc_activate()`, 
which are called internally.}

\item{file_rmd}{String. Name of RMarkdown file to render. Default is "R-validation.Rmd".
Also included is "R-validationPreamble.Rmd", which contains the preamble and introduction 
portion of the report. The RMarkdown file should be located in the `inst/qualify_r` 
folder of the package.}

\item{verbose}{Logical. If TRUE, prints progress messages to the console.}
}
\value{
The path to the `R-validation` folder. The primary purpose of this 
  function is its side effects, rendering an RMarkdown document.
}
\description{
Run IQ-OQ on an installation of R software
}
\details{
This function creates a folder named R-validation at the specified path,
  sets up the environment, and executes the validation process.
  The output of the validation will be saved in the created folder. 

The validation process involves running a series of tests on the R installation and
can be quite time consuming. The function will print progress messages to the 
console if `verbose` is set to TRUE.

The following steps are carried out:

\enumerate{
  \item Create the folder tree R-validation/IQ-OQ-TestOutput at path_save, or selected folder
  \item Install TinyTeX and necessary LaTeX packages
  \item Install Pandoc
  \item Copy Rmd file to the R-validation folder
  \item Execute the IQ-OQ by rendering the Rmd to LaTeX
  \item Compile the LaTeX file to pdf
}
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\donttest{
rqualify(path_save     = tempdir(),
         setup_tinytex = TRUE,
         setup_pandoc  = TRUE,
         file_rmd      = "R-validationPreamble.Rmd")
}
\dontshow{\}) # examplesIf}
}
