% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqualify.R
\name{rqualify}
\alias{rqualify}
\title{Run IQ-OQ on an installation of R software}
\usage{
rqualify(
  path_save,
  setup_tinytex = TRUE,
  setup_pandoc = TRUE,
  verbose = TRUE,
  create_zip = TRUE
)
}
\arguments{
\item{path_save}{Character. Path to save the R-validation folder. Ensure
a folder named `R-validation` does not already exist at this location.}

\item{setup_tinytex}{Logical. If TRUE, sets up tinytex for LaTeX document
generation.}

\item{setup_pandoc}{Logical. If TRUE, sets up pandoc for document conversion.}

\item{verbose}{Logical. If TRUE, prints progress messages to the console.}

\item{create_zip}{Logical. If TRUE, creates a zip file of the R-validation folder after validation.}
}
\description{
Run IQ-OQ on an installation of R software
}
\details{
This function creates a folder named R-validation at the specified path,
sets up the environment, and executes the validation process.
The output of the validation will be saved in the created folder. By default,
the contents of the R-validation folder will be zipped up if the validation
is successful.

The following steps are carried out:

\enumerate{
  \item Create the folder tree R-validation/IQ-OQ-TestOutput at path_save, or selected folder
  \item Install tinytex and necessary LaTeX packages
  \item Install pandoc
  \item Copy R-validation.Rmd to the R-validation folder
  \item Execute the IQ-OQ by rendering R-validation.Rmd to R-validation.tex
  \item Compile R-validation.tex to R-validation.pdf
  \item If IQ-OQ is successful, create R-validation.zip at folder where R-validation is located
}
}
\examples{
\dontrun{
rqualify(path_save  = path.expand('~'))
}

}
