% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqualify_setup.R
\name{rqualify_setup}
\alias{rqualify_setup}
\title{Initialize environment for IQ-OQ.}
\usage{
rqualify_setup(
  path_save,
  setup_tinytex = TRUE,
  setup_pandoc = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{path_save}{Character. Path where the 'R-validation' folder will be created
to store validation outputs.}

\item{setup_tinytex}{Logical. If TRUE, sets up tinytex for LaTeX document
generation.}

\item{setup_pandoc}{Logical. If TRUE, sets up pandoc for document conversion.}

\item{verbose}{Logical. If TRUE, prints progress messages to the console.}
}
\description{
Initialize environment for IQ-OQ.
}
\details{
The function sets up the environment by configuring locale settings and
  installing any dependencies, including tinytex and pandoc.
}
\examples{
\dontrun{
rqualify_setup(path_save = path.expand("~"))
}

}
\seealso{
\code{\link{rqualify}} which calls this function internally.
}
